+++ bin/screensaver_manager
@@ -10,7 +10,7 @@ sys.path.append(root_dir)
 import argparse
 from pifi.config import Config
 from pifi.logger import Logger
-from pifi.screensavermanager import ScreensaverManager
+from pifi.screensaver.screensavermanager import ScreensaverManager
 
 def parseArgs():
     parser = argparse.ArgumentParser(
